fix(storybook): fix initial color scheme
