+++ packages/storybook/.storybook/backgroundOptions.ts
@@ -1,0 +1,6 @@
+export const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+
+export const backgroundOptions = {
+    dark: {name: 'Dark', value: 'var(--coveo-app-background)'},
+    light: {name: 'Light', value: 'var(--coveo-app-background)'},
+};
